Support batch task for Inlong group
