+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sort/CreateSortConfigListener.java
@@ -18,96 +18,195 @@
 package org.apache.inlong.manager.service.sort;
 
 import com.google.common.collect.Lists;
-import org.apache.commons.collections.CollectionUtils;
+import com.google.common.collect.Maps;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.inlong.common.enums.DataTypeEnum;
 import org.apache.inlong.manager.common.consts.InlongConstants;
+import org.apache.inlong.manager.common.enums.ClusterType;
 import org.apache.inlong.manager.common.enums.GroupOperateType;
+import org.apache.inlong.manager.common.enums.MQType;
+import org.apache.inlong.manager.common.enums.SourceType;
 import org.apache.inlong.manager.common.exceptions.WorkflowListenerException;
+import org.apache.inlong.manager.common.pojo.cluster.ClusterInfo;
+import org.apache.inlong.manager.common.pojo.cluster.pulsar.PulsarClusterInfo;
 import org.apache.inlong.manager.common.pojo.group.InlongGroupExtInfo;
 import org.apache.inlong.manager.common.pojo.group.InlongGroupInfo;
 import org.apache.inlong.manager.common.pojo.sink.StreamSink;
+import org.apache.inlong.manager.common.pojo.source.StreamSource;
+import org.apache.inlong.manager.common.pojo.source.kafka.KafkaSource;
+import org.apache.inlong.manager.common.pojo.source.pulsar.PulsarSource;
+import org.apache.inlong.manager.common.pojo.stream.InlongStreamInfo;
 import org.apache.inlong.manager.common.pojo.workflow.form.process.GroupResourceProcessForm;
+import org.apache.inlong.manager.service.cluster.InlongClusterService;
 import org.apache.inlong.manager.service.sink.StreamSinkService;
-import org.apache.inlong.manager.service.sort.util.DataFlowUtils;
+import org.apache.inlong.manager.service.sort.util.ExtractNodeUtils;
+import org.apache.inlong.manager.service.sort.util.LoadNodeUtils;
+import org.apache.inlong.manager.service.source.StreamSourceService;
 import org.apache.inlong.manager.workflow.WorkflowContext;
 import org.apache.inlong.manager.workflow.event.ListenerResult;
 import org.apache.inlong.manager.workflow.event.task.SortOperateListener;
 import org.apache.inlong.manager.workflow.event.task.TaskEvent;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.inlong.sort.protocol.GroupInfo;
+import org.apache.inlong.sort.protocol.StreamInfo;
+import org.apache.inlong.sort.protocol.enums.PulsarScanStartupMode;
+import org.apache.inlong.sort.protocol.node.Node;
+import org.apache.inlong.sort.protocol.transformation.relation.NodeRelation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
-/**
- * Create sort config when disable the ZooKeeper
- */
-@Deprecated
 @Component
+@Slf4j
 public class CreateSortConfigListener implements SortOperateListener {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(CreateSortConfigListener.class);
-    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper(); // thread safe
+    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
     @Autowired
-    private StreamSinkService streamSinkService;
+    private StreamSourceService sourceService;
     @Autowired
-    private DataFlowUtils dataFlowUtils;
+    private StreamSinkService sinkService;
+    @Autowired
+    private InlongClusterService clusterService;
 
     @Override
     public TaskEvent event() {
         return TaskEvent.COMPLETE;
     }
 
+    @SneakyThrows
     @Override
-    public ListenerResult listen(WorkflowContext context) throws WorkflowListenerException {
-        LOGGER.info("Create sort config for groupId={}", context.getProcessForm().getInlongGroupId());
+    public ListenerResult listen(WorkflowContext context) {
+        log.info("Create sort config V2 for groupId={}", context.getProcessForm().getInlongGroupId());
         GroupResourceProcessForm form = (GroupResourceProcessForm) context.getProcessForm();
         GroupOperateType groupOperateType = form.getGroupOperateType();
         if (groupOperateType == GroupOperateType.SUSPEND || groupOperateType == GroupOperateType.DELETE) {
             return ListenerResult.success();
         }
         InlongGroupInfo groupInfo = form.getGroupInfo();
-        String groupId = groupInfo.getInlongGroupId();
-        if (StringUtils.isEmpty(groupId)) {
-            LOGGER.warn("GroupId is null for context={}", context);
+        List<InlongStreamInfo> streamInfos = form.getStreamInfos();
+        int sinkCount = streamInfos.stream()
+                .map(s -> s.getSinkList() == null ? 0 : s.getSinkList().size())
+                .reduce(0, Integer::sum);
+        if (sinkCount == 0) {
+            log.warn("not any sink for group {} found, skip creating sort config", groupInfo.getInlongGroupId());
             return ListenerResult.success();
         }
 
-        List<StreamSink> streamSinks = streamSinkService.listSink(groupId, null);
-        if (CollectionUtils.isEmpty(streamSinks)) {
-            LOGGER.warn("Sink not found by groupId={}", groupId);
-            return ListenerResult.success();
-        }
-
-        try {
-            // TODO Support more than one sinks under a stream
-            // Map<String, DataFlowInfo> dataFlowInfoMap = streamSinks.stream().map(sink -> {
-            //             DataFlowInfo flowInfo = dataFlowUtils.createDataFlow(groupInfo, sink);
-            //             return Pair.of(sink.getInlongStreamId(), flowInfo);
-            //         }
-            // ).collect(Collectors.toMap(Pair::getKey, Pair::getValue));
-
-            // String dataFlows = OBJECT_MAPPER.writeValueAsString(dataFlowInfoMap);
-            InlongGroupExtInfo extInfo = new InlongGroupExtInfo();
-            extInfo.setInlongGroupId(groupId);
-            extInfo.setKeyName(InlongConstants.DATA_FLOW);
-            // extInfo.setKeyValue(dataFlows);
-            if (groupInfo.getExtList() == null) {
-                groupInfo.setExtList(Lists.newArrayList());
-            }
-            upsertDataFlow(groupInfo, extInfo);
-        } catch (Exception e) {
-            LOGGER.error("create sort config failed for sink list={} ", streamSinks, e);
-            throw new WorkflowListenerException("create sort config failed: " + e.getMessage());
-        }
+        GroupInfo configInfo = createGroupInfo(groupInfo, streamInfos);
+        String dataFlows = OBJECT_MAPPER.writeValueAsString(configInfo);
+        addExtInfo(groupInfo, InlongConstants.DATA_FLOW, dataFlows);
         return ListenerResult.success();
     }
 
-    private void upsertDataFlow(InlongGroupInfo groupInfo, InlongGroupExtInfo extInfo) {
-        groupInfo.getExtList().removeIf(ext -> InlongConstants.DATA_FLOW.equals(ext.getKeyName()));
+    private void addExtInfo(InlongGroupInfo groupInfo, String key, String value) {
+        if (groupInfo.getExtList() == null) {
+            groupInfo.setExtList(Lists.newArrayList());
+        }
+        InlongGroupExtInfo extInfo = new InlongGroupExtInfo();
+        extInfo.setInlongGroupId(groupInfo.getInlongGroupId());
+        extInfo.setKeyName(key);
+        extInfo.setKeyValue(value);
+        upsertExtInfo(groupInfo, extInfo);
+    }
+
+    /**
+     * TODO need support TubeMQ
+     */
+    private GroupInfo createGroupInfo(InlongGroupInfo groupInfo, List<InlongStreamInfo> streamInfoList) {
+        String groupId = groupInfo.getInlongGroupId();
+        List<StreamSink> streamSinks = sinkService.listSink(groupId, null);
+        Map<String, List<StreamSink>> sinkMap = streamSinks.stream()
+                .collect(Collectors.groupingBy(StreamSink::getInlongStreamId, HashMap::new,
+                        Collectors.toCollection(ArrayList::new)));
+        Map<String, List<StreamSource>> sourceMap = createPulsarSources(groupInfo, streamInfoList);
+
+        List<StreamInfo> streamInfos = new ArrayList<>();
+        for (InlongStreamInfo inlongStream : streamInfoList) {
+            String streamId = inlongStream.getInlongStreamId();
+            StreamInfo streamInfo = new StreamInfo(streamId,
+                    createNodesForStream(sourceMap.get(streamId), sinkMap.get(streamId)),
+                    createNodeRelationsForStream(sourceMap.get(streamId), sinkMap.get(streamId)));
+            streamInfos.add(streamInfo);
+        }
+
+        return new GroupInfo(groupId, streamInfos);
+    }
+
+    private Map<String, List<StreamSource>> createPulsarSources(
+            InlongGroupInfo groupInfo, List<InlongStreamInfo> streamInfoList) {
+
+        if (!MQType.MQ_PULSAR.equals(groupInfo.getMqType())) {
+            String errMsg = String.format("Unsupported MQ type %s", groupInfo.getMqType());
+            log.error(errMsg);
+            throw new WorkflowListenerException(errMsg);
+        }
+
+        Map<String, List<StreamSource>> sourceMap = Maps.newHashMap();
+        ClusterInfo clusterInfo = clusterService.getOne(groupInfo.getInlongClusterTag(), null,
+                ClusterType.PULSAR);
+
+        PulsarClusterInfo pulsarCluster = (PulsarClusterInfo) clusterInfo;
+        String adminUrl = pulsarCluster.getAdminUrl();
+        String serviceUrl = pulsarCluster.getUrl();
+        String tenant = StringUtils.isEmpty(pulsarCluster.getTenant()) ? InlongConstants.DEFAULT_PULSAR_TENANT
+                : pulsarCluster.getTenant();
+        streamInfoList.forEach(streamInfo -> {
+            PulsarSource pulsarSource = new PulsarSource();
+            String streamId = streamInfo.getInlongStreamId();
+            pulsarSource.setTenant(tenant);
+            pulsarSource.setSourceName(streamId);
+            pulsarSource.setNamespace(groupInfo.getMqResource());
+            pulsarSource.setTopic(streamInfo.getMqResource());
+            pulsarSource.setAdminUrl(adminUrl);
+            pulsarSource.setServiceUrl(serviceUrl);
+            pulsarSource.setInlongComponent(true);
+
+            List<StreamSource> sourceInfos = sourceService.listSource(groupInfo.getInlongGroupId(), streamId);
+            for (StreamSource sourceInfo : sourceInfos) {
+                if (StringUtils.isEmpty(pulsarSource.getSerializationType())
+                        && StringUtils.isNotEmpty(sourceInfo.getSerializationType())) {
+                    pulsarSource.setSerializationType(sourceInfo.getSerializationType());
+                }
+                if (SourceType.forType(sourceInfo.getSourceType()) == SourceType.KAFKA) {
+                    pulsarSource.setPrimaryKey(((KafkaSource) sourceInfo).getPrimaryKey());
+                }
+            }
+            if (StringUtils.isEmpty(pulsarSource.getSerializationType())) {
+                pulsarSource.setSerializationType(DataTypeEnum.CSV.getName());
+            }
+            pulsarSource.setScanStartupMode(PulsarScanStartupMode.EARLIEST.getValue());
+            pulsarSource.setFieldList(streamInfo.getFieldList());
+            sourceMap.computeIfAbsent(streamId, key -> Lists.newArrayList()).add(pulsarSource);
+        });
+        return sourceMap;
+    }
+
+    private List<Node> createNodesForStream(List<StreamSource> sources, List<StreamSink> streamSinks) {
+        List<Node> nodes = Lists.newArrayList();
+        nodes.addAll(ExtractNodeUtils.createExtractNodes(sources));
+        nodes.addAll(LoadNodeUtils.createLoadNodes(streamSinks));
+        return nodes;
+    }
+
+    private List<NodeRelation> createNodeRelationsForStream(List<StreamSource> sources, List<StreamSink> streamSinks) {
+        NodeRelation relation = new NodeRelation();
+        List<String> inputs = sources.stream().map(StreamSource::getSourceName).collect(Collectors.toList());
+        List<String> outputs = streamSinks.stream().map(StreamSink::getSinkName).collect(Collectors.toList());
+        relation.setInputs(inputs);
+        relation.setOutputs(outputs);
+        return Lists.newArrayList(relation);
+    }
+
+    private void upsertExtInfo(InlongGroupInfo groupInfo, InlongGroupExtInfo extInfo) {
+        groupInfo.getExtList().removeIf(ext -> extInfo.getKeyName().equals(ext.getKeyName()));
         groupInfo.getExtList().add(extInfo);
     }
