+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/workflow/listener/StreamTaskListenerFactory.java
@@ -29,7 +29,7 @@ import org.apache.inlong.manager.service.mq.PulsarTopicCreateSelector;
 import org.apache.inlong.manager.service.mq.PulsarTopicDeleteSelector;
 import org.apache.inlong.manager.service.resource.StreamSinkResourceListener;
 import org.apache.inlong.manager.service.sort.CreateStreamSortConfigListener;
-import org.apache.inlong.manager.service.sort.ZookeeperEnabledSelector;
+import org.apache.inlong.manager.service.sort.ZookeeperDisabledSelector;
 import org.apache.inlong.manager.workflow.WorkflowContext;
 import org.apache.inlong.manager.workflow.definition.ServiceTaskListenerProvider;
 import org.apache.inlong.manager.workflow.definition.ServiceTaskType;
@@ -81,7 +81,7 @@ public class StreamTaskListenerFactory implements PluginBinder, ServiceTaskListe
         queueOperateListeners.put(createPulsarSubscriptionTaskListener, new PulsarTopicCreateSelector());
         queueOperateListeners.put(deletePulsarTopicTaskListener, new PulsarTopicDeleteSelector());
         sortOperateListeners = new LinkedHashMap<>();
-        sortOperateListeners.put(createSortConfigListener, new ZookeeperEnabledSelector());
+        sortOperateListeners.put(createSortConfigListener, new ZookeeperDisabledSelector());
         sinkOperateListeners = new LinkedHashMap<>();
         sinkOperateListeners.put(sinkResourceListener, context -> {
             ProcessForm processForm = context.getProcessForm();
